moved tasks funcs to repo for smooth upd
